requirements.txt
pyotp==2.6.0
playwright==1.12.0
requests==2.25.1
pytest==6.2.4
pytest-base-url==1.4.2
pytest-bdd==4.0.2
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-html-reporter==0.2.6
pytest-metadata==1.11.0
pytest-order==1.0.0
pytest-playwright==0.1.2
pytest-variables==1.9.0
pytest-xdist==2.3.0
python-dateutil==2.7.3
python-dotenv==0.17.1visu
python-slugify==5.0.2
pytest-rerunfailures==10.1
pyOpenSSL==20.0.1
filelock==3.0.12
